Modified interface to separate 'build' from 'from_dict'
This clarifies the testing boundary a litte, as the 'from_dict' test
becomes about wrangling the dictionary into the correct form for calling
the 'build' interface. It also removes the need to create dictionaries
in order to test (i.e. no need to create an 'encoded actions string' to
test the build behaivour).