Issue #1 Fix
Adding new 'Update with External Reference Action'

These changes add a new action that is more explicit about performing
model object updates and then managing the ExternalKeyMapping object
that is related to the model object.

These changes are part of the work to remove the
'AlignExternalReferenceAction' class. They are necessary because there
is a bug with updating objects that retain their external system keys
but change their match fields.