Removing buggy CSV file checks
The 'has_header()' check has been extremely buggy. In particular, I
found that the change from 'match_field_name' to 'match_field_names'
caused this to have a problem for a file I was syncing from (the field I
was specifying was also 17 characters in string length, which causes the
logic at the end of the 'has_header()' function in csv.py to fail).

I was also not testing this...