Allowing 'match_on' fields to be OR'd
This is the first implementation of allowing logical operators to be
used in the match_on field. The proper implementation will come shortly
and should use a postfix style of notation.

For now, the match_on fields are either all ANDed, the default, or all
Red, if a '|' is the last character.